+++ src/components/legend/legend-list/legend-item/legend-item-timeline/styles.scss
@@ -1,0 +1,18 @@
+@import 'settings';
+@import 'input-range';
+
+.c-legend-timeline {
+  display: flex;
+  margin: $space-1 * 3 $space-1 * 3 $space-1 * 2 0;
+
+  > button {
+    padding: 0;
+    cursor: pointer;
+  }
+
+  .timeline-play-button {
+    border: none;
+    fill: $color-dark-1;
+    margin-right: 12px;
+  }
+}
