+++ src/css/input-range.scss
@@ -1,0 +1,36 @@
+// Input range styles
+:global {
+  @import 'react-input-range/src/scss/index.scss';
+
+  .input-range__label {
+    font-size: 8px;
+    color: $color-dark-2;
+    top: -24px;
+  }
+
+  .input-range__slider {
+    border: 0;
+    background-color: $color-primary;
+    width: 14px;
+    height: 14px;
+    margin-top: -9px;
+  }
+
+  .input-range__track {
+    height: 3px;
+  }
+
+  .input-range__track--active {
+    background-color: $color-primary;
+  }
+
+  .input-range__label--min, .input-range__label--max {
+    font-size: 10px;
+    top: auto;
+    bottom: -16px;
+
+    .input-range__label-container {
+      left: 0;
+    }
+  }
+}
