+++ src/css/settings.scss
@@ -30,3 +30,7 @@ $font-size-h4: 16px;
 $font-weight-light: 300;
 $font-weight-regular: 400;
 $font-weight-bold: 700;
+
+
+// Space
+$space-1: 8px;
